@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'game_view.dart';
 
 class HomeView extends StatelessWidget {
-  const HomeView({Key? key}) : super(key: key);
+  const HomeView({super.key});
 
   @override
   Widget build(BuildContext context) {
